+++ src/main/java/br/com/fiap/on/smarthas/model/repositories/PerfilRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface PerfilRepository extends JpaRepository<PerfilORM, Integer> {
+public interface PerfilRepository extends JpaRepository<PerfilORM, Long> {
     List<PerfilORM> findByNomeContaining(String nome);
     Optional<PerfilORM> findByMnemonico(String mnemonico);
 }
