+++ src/main/java/br/com/fiap/on/smarthas/model/repositories/PermissaoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface PermissaoRepository extends JpaRepository<PermissaoORM, Integer> {
+public interface PermissaoRepository extends JpaRepository<PermissaoORM, Long> {
     Optional<PermissaoORM> findByNome(String nome);
 }
