+++ src/main/java/br/com/fiap/on/smarthas/model/repositories/UsuarioPerfilRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface UsuarioPerfilRepository extends JpaRepository<UsuarioPerfilORM, Integer> {
+public interface UsuarioPerfilRepository extends JpaRepository<UsuarioPerfilORM, Long> {
     List<UsuarioPerfilORM> findByUsuario(UsuarioORM usuario);
     List<UsuarioPerfilORM> findByPerfil(PerfilORM perfil);
 }
